pyproject.toml
[project]
name = "marketplace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "granian[extras]>=1.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.10.0",
    "google-auth[requests]>=2.37.0",
    "requests>=2.32.0",
    "httpx>=0.28.0",
    "redis[hiredis]>=5.2.0",
    "arq>=0.26.0",
    "slowapi>=0.1.9",
    "structlog>=25.1.0",
    "python-multipart>=0.0.20",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "boto3>=1.36.0",
    "email-validator>=2.2.0",
    "phonenumbers>=8.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
